pyproject.toml
[project]
name = "aire-health-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.124.0",
    "langchain>=1.1.2",
    "langchain-openai>=1.1.1",
    "langgraph>=1.0.4",
    "mysql-connector-python>=9.5.0",
    "openai>=2.9.0",
    "pathlib>=1.0.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "qdrant-client[fastembed]>=1.16.1",
    "sqlalchemy>=2.0.44",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
]
